% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-massbank.R
\name{metaDataBlocks}
\alias{metaDataBlocks}
\title{Metadata blocks to be read}
\usage{
metaDataBlocks()
}
\value{
A \code{data.frame} with metadata blocks.
}
\description{
\code{metaDataBlocks} returns a \code{data.frame} with the MassBank metadata blocks
and whether they should be imported by default from the MassBank text files.
}
\examples{

metaDataBlocks()
}
\author{
Michael Witting
}
