% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-massbank.R
\name{metaDataBlocks}
\alias{metaDataBlocks}
\title{Metadata blocks to be read}
\usage{
metaDataBlocks()
}
\value{
A \code{data.frame} with metadata blocks.
}
\description{
It is possible to
}
\examples{
metaDataBlocks()
}
\author{
Michael Witting
}
